+++ src/booklab/my_books/books_manager.py
@@ -22,16 +22,17 @@ class MyBook:
 
     **Mandatory requirements:**
 
-    - any Jinja renderings will be made "from string" (ie, ising Flask render_from_string() which is included) or by creating a local Jinja environment.
+    - any Jinja renderings will be made "from string" (ie, using 
+    Flask `render_from_string()` which is included) or by creating a local Jinja environment.
 
     author: Petre Iordanescu (petre.iordanescu@gmail.com)
     """
-    _MY_BOOKS_URL_prefix = "/my-books/"
-    MY_BOOK_URL: str
-    MY_BOOKS_PATH: str = MY_BOOKS_ROOT
-    book_code: str
-    db_books_catalog: pysondb
-    db_book_nav: pysondb
+    MY_BOOKS_URL_prefix: str = "/my-books/"  # URL prefix to add when accesing a book local (generated) site
+    MY_BOOK_URL: str = None  # instantiated book URL to local (generated) site
+    MY_BOOKS_ROOT: str = MY_BOOKS_ROOT  # confusing name ? just duplicate the global one in class namespace
+    book_code: str = None  # instanciated book code
+    db_books_catalog: pysondb = None  # books catalog data controller
+    db_book_nav: pysondb = None  # books navigation data controller
 
     def __init__(
         self,
@@ -40,13 +41,13 @@ class MyBook:
     ):
         """Init an instance of class MyBook
         """
-        self.MY_BOOKS_PATH = MY_BOOKS_ROOT
+        self.MY_BOOKS_ROOT = MY_BOOKS_ROOT  # confusing name ? just duplicate the global one in class namespace
         self.book_code = book_code
         self.db_books_catalog = db
         self.MY_BOOK_URL = w3lib.url.canonicalize_url(
             url_quote(
                 str(FULL_EXT_URL) +
-                str(MyBook._MY_BOOKS_URL_prefix) +
+                str(MyBook.MY_BOOKS_URL_prefix) +
                 str(self.book_code) +
                 "/docs/"
             )
@@ -178,7 +179,7 @@ class MyBook:
         """
         my_book_path = os.path.abspath(
             os.path.join(
-                self.MY_BOOKS_PATH,
+                self.MY_BOOKS_ROOT,
                 self.book_code
             )
         )