+++ CHANGELOG.md
@@ -27,6 +27,8 @@ Implement `bbld` and update `book_template`:
 * ... [0.10a___] `MyBook.renderBookConfig()` steo 2
     - ...
     - ... read mkdocs_template.yml to render from string
+    - ... rename `MyBook` to `MyBooks`
+    - [0.10.dev17] `MyBook` rename MY_BOOKS_PATH attribute to MY_BOOKS_ROOT and comment all class variables
     - [0.10.dev16] `MyBook` clean MY_BOOKS_PATH attribute and make it equal to MY_BOOKS_ROOT created on booklabd init
     - [0.10.dev15] `MyBook` drop "book_path" attribute because exists mwthod `getBookPath()` that do that
 * [0.10a14] upd README.md with pj GitHub